+++ package.json
@@ -7,9 +7,7 @@
   "scripts": {
     "build:lib": "babel src --out-dir lib",
     "build":
-      "npm run build:lib && NODE_ENV=production webpack --config webpack.prod.config.js && cp src/react-tether.d.ts lib/react-tether.d.ts",
-    "postbuild":
-      "NODE_ENV=production TARGET=minify webpack --config webpack.prod.config.js",
+      "npm run build:lib && webpack --mode production && cp src/react-tether.d.ts lib/react-tether.d.ts",
     "prebuild": "rm -rf dist && mkdir dist && rm -rf lib && mkdir lib",
     "prepublish": "npm run build",
     "demo": "parcel example/index.html",
@@ -84,7 +82,8 @@
     "testcafe": "0.19.2",
     "testcafe-browser-provider-saucelabs": "1.3.0",
     "typescript": "2.8.3",
-    "webpack": "1.15.0",
+    "webpack": "4.6.0",
+    "webpack-cli": "2.0.15",
     "xo": "0.20.3"
   },
   "prettier": {
